Move os module
