+++ llrt_core/src/modules/mod.rs
@@ -1,4 +1,4 @@
-pub use llrt_modules::{buffer, fs, path};
+pub use llrt_modules::{buffer, fs, os, path};
 
 pub mod child_process;
 pub mod console;
@@ -11,7 +11,6 @@ pub mod llrt;
 pub mod module;
 pub mod navigator;
 pub mod net;
-pub mod os;
 pub mod performance;
 pub mod process;
 pub mod timers;
