+++ llrt_modules/Cargo.toml
@@ -10,11 +10,13 @@ all = ["buffer", "fs", "os", "path"]
 
 buffer = ["llrt_utils/encoding"]
 fs = ["tokio/fs", "llrt_utils/fs", "ring", "buffer", "path"]
-os = ["libc"]
+os = ["process", "libc", "once_cell"]
 path = []
+process = []
 
 [dependencies]
 either = "1"
+once_cell = { version = "1", optional = true }
 llrt_utils = { path = "../llrt_utils", default-features = false }
 rquickjs = { version = "0.6", features = [
   "either",
